+++ packages/host/src/export.ts
@@ -245,6 +245,18 @@ async function loadLayeredIcon(
 	return icon;
 }
 
+function fixIcons(item: ItemPrototype) {
+	const icons = item.icons;
+	if (typeof icons === "object" && !(icons instanceof Array) && icons !== null) {
+		// It's possible to specify icons as an object with arbitrary keys
+		// and the game will still accept it, cast the value to array if
+		// this is the case.
+		item = { ...item };
+		item.icons = Object.values(icons);
+	}
+	return item;
+}
+
 const itemTypes = new Set([
 	"item",
 	"ammo",
@@ -277,7 +289,8 @@ const itemTypes = new Set([
 function filterItems(prototypes: Prototypes): ItemPrototype[] {
 	return Object.entries(prototypes)
 		.filter(([type, _]) => itemTypes.has(type))
-		.flatMap(([_, typePrototypes]) => Object.values(typePrototypes)) as ItemPrototype[]
+		.flatMap(([_, typePrototypes]) => Object.values(typePrototypes) as ItemPrototype[])
+		.map(fixIcons)
 	;
 }