+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed Research Sync applying progress of a previous techology level to the level on the controller when it's the active research during instance startup.
 - Fixed data export with a mod using shorthand notation for the default value of a color-setting causing the ModPack data to be set to invalid data.
 - Fixed per user per instance player stats not being deleted when the instance it's for is deleted.
+- Fixed mods setting the item prototype icons property to an object mapping instead of an array breaking the data export.
 
 
 ## Version 2.0.0-alpha.16